worker/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
playwright>=1.40.0
aiohttp>=3.9.0
supabase>=2.4.0
pydantic>=2.0.0
rendercv>=2.3
ruamel.yaml>=0.18.0
pypdf>=5.0.0
python-multipart>=0.0.9
